refactor: 예외 메시지 처리 구조 수정
