+++ src/main/java/cart/domain/product/ProductService.java
@@ -1,8 +1,8 @@
 package cart.domain.product;
 
+import cart.exception.ErrorCode;
+import cart.exception.GlobalException;
 import cart.web.controller.product.dto.ProductRequest;
-import cart.web.exception.ErrorCode;
-import cart.web.exception.GlobalException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Service;
