+++ src/main/java/cart/exception/ErrorCode.java
@@ -1,0 +1,10 @@
+package cart.exception;
+
+public enum ErrorCode {
+    INVALID_CATEGORY,
+    PRODUCT_NOT_FOUND,
+    INVALID_DELETE,
+    USER_NOT_FOUND,
+    INVALID_REQUEST,
+    UNEXPECTED_EXCEPTION;
+}
