+++ src/main/java/cart/web/handler/GlobalExceptionHandler.java
@@ -1,8 +1,8 @@
 package cart.web.handler;
 
-import cart.web.exception.ErrorCode;
-import cart.web.exception.ErrorResponse;
-import cart.web.exception.GlobalException;
+import cart.exception.ErrorCode;
+import cart.exception.ErrorResponse;
+import cart.exception.GlobalException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.context.support.DefaultMessageSourceResolvable;
@@ -12,18 +12,30 @@ import org.springframework.web.bind.MethodArgumentNotValidException;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
 
+import java.util.EnumMap;
 import java.util.List;
+import java.util.Map;
 import java.util.stream.Collectors;
 
 @RestControllerAdvice
 public class GlobalExceptionHandler {
 
     public final Logger log = LoggerFactory.getLogger(getClass());
+    private final Map<ErrorCode, String> errorEntry = new EnumMap<>(ErrorCode.class);
+
+    {
+        errorEntry.put(ErrorCode.INVALID_CATEGORY, "상품 정보를 찾을 수 없습니다.");
+        errorEntry.put(ErrorCode.PRODUCT_NOT_FOUND, "삭제할 수 없습니다. 관리자에게 문의하세요.");
+        errorEntry.put(ErrorCode.INVALID_DELETE, "유효하지 않은 카테고리입니다.");
+        errorEntry.put(ErrorCode.USER_NOT_FOUND, "해당 유저가 존재하지 않습니다.");
+        errorEntry.put(ErrorCode.INVALID_REQUEST, "");
+        errorEntry.put(ErrorCode.UNEXPECTED_EXCEPTION, "예상치 못한 예외가 발생했습니다. 잠시만 기다려주세요.");
+    }
 
     @ExceptionHandler(GlobalException.class)
     public ResponseEntity<ErrorResponse> globalException(final GlobalException e) {
         final ErrorCode errorCode = e.getErrorCode();
-        final ErrorResponse errorResponse = new ErrorResponse(errorCode, List.of(errorCode.getMessage()));
+        final ErrorResponse errorResponse = new ErrorResponse(errorCode, List.of(errorEntry.get(errorCode)));
         return new ResponseEntity<>(errorResponse, HttpStatus.BAD_REQUEST);
     }
