+++ src/test/java/cart/service/ProductServiceTest.java
@@ -4,8 +4,8 @@ import cart.dao.ProductDaoImpl;
 import cart.domain.product.Product;
 import cart.domain.product.ProductCategory;
 import cart.domain.product.ProductService;
+import cart.exception.GlobalException;
 import cart.web.controller.product.dto.ProductRequest;
-import cart.web.exception.GlobalException;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
